feat: add usersListReducer and logic for it
